refactor(metrics): cleanup
- remove LegacyField
- move Type to a dedicated file
